pyproject.toml
[project]
name = "pyGandalf"
version = "0.0.1"
description = "Default template for PDM package"
authors = [
    {name = "George Papagiannakis", email = ""},
    {name = "John Petropoulos", email = "johnaipeia@gmail.com"},
    {name = "Stratos Geronikolakis", email = ""},
    {name = "Manos Kamarianakis", email = ""},
    {name = "Antonis Protopsaltis", email = ""},
]
dependencies = [
    "numpy>=1.26.4",
    "glfw>=2.7.0",
    "PyOpenGL>=3.1.7",
    "pillow>=10.2.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "Apache-2.0"}


[tool.pdm]
distribution = false
